Harden upload/delete handling and replace base64 with AES encryption

+++ delete.php
@@ -1,25 +1,26 @@
 <?php
+
+if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
+	http_response_code(405);
+	exit('Method not allowed.');
+}
+
 $up_folder_path = "upload";
 $down_folder_path = "download";
 
-// List of name of files inside
-$upfiles = glob($up_folder_path.'/*');
-$downfiles = glob($down_folder_path.'/*');
+$upfiles = is_dir($up_folder_path) ? glob($up_folder_path . '/*') : [];
+$downfiles = is_dir($down_folder_path) ? glob($down_folder_path . '/*') : [];
 
-// Deleting all the files in the list
-foreach($upfiles as $file) {
-
-	if(is_file($file))
-		// Delete the given file
+foreach ($upfiles as $file) {
+	if (is_file($file)) {
 		unlink($file);
+	}
 }
-foreach($downfiles as $file) {
-
-	if(is_file($file))
-		// Delete the given file
+foreach ($downfiles as $file) {
+	if (is_file($file)) {
 		unlink($file);
+	}
 }
-print("All Files are Deleted Successfully.");
 
+print("All files deleted successfully.");
 echo "<meta http-equiv='refresh' content='2;url=index.php'/>";
-?>
